Finish add usecase

+++ cnccode/customClasses/AdditionalChargesRates/Infra/Persistence/AdditionalChargeRatePDORepository.php
@@ -49,21 +49,27 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
     function save(AdditionalChargeRate $additionalChargeRate)
     {
         $this->pdo->beginTransaction();
-        $this->deleteAdditionalChargeCustomerSpecificPrices($additionalChargeRate);
-        $this->insertAdditionalChargeCustomerSpecificPrices($additionalChargeRate);
-        $query                   = "insert into additionalChargeRate(id,description,notes,salesPrice) values (?,?,?,?) on duplicate key update description = ?, notes = ?, salesPrice = ?";
-        $insertOrUpdateStatement = $this->pdo->prepare($query);
-        if (!$insertOrUpdateStatement->execute(
-            [
-                $additionalChargeRate->id()->value(),
-                $additionalChargeRate->description()->value(),
-                $additionalChargeRate->notes()->value(),
-                $additionalChargeRate->salePrice()->value(),
-            ]
-        )) {
+        try {
+            $query                   = "insert into additionalChargeRate(id,description,notes,salesPrice) values (:id,:description,:notes,:salesPrice) on duplicate key update description = :description, notes = :notes, salesPrice = :salesPrice";
+            $insertOrUpdateStatement = $this->pdo->prepare($query);
+            if (!$insertOrUpdateStatement->execute(
+                [
+                    "id"          => $additionalChargeRate->id()->value(),
+                    "description" => $additionalChargeRate->description()->value(),
+                    "notes"       => $additionalChargeRate->notes()->value(),
+                    "salesPrice"  => $additionalChargeRate->salePrice()->value(),
+                ]
+            )) {
+                throw new PDOException('Failed to insert or update');
+            }
+            $this->deleteAdditionalChargeCustomerSpecificPrices($additionalChargeRate);
+            $this->insertAdditionalChargeCustomerSpecificPrices($additionalChargeRate);
+            $this->pdo->commit();
+        } catch (\Exception $exception) {
             $this->pdo->rollBack();
+            error_log("Failed to insert or update Additional Charge Rate: {$exception->getMessage()}");
+            throw new PDOException("Failed to insert or update: {$exception->getMessage()}");
         }
-        $this->pdo->commit();
     }
 
     /**
@@ -72,9 +78,13 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
     private function deleteAdditionalChargeCustomerSpecificPrices(AdditionalChargeRate $additionalChargeRate): void
     {
         $deleteAdditionalChargeCustomerPricesStatement = $this->pdo->prepare(
-            'delete * from additionalChargeRateCustomerPrices where additionalChargeRateId = ?'
+            'delete from additionalChargeRateCustomerPrices where additionalChargeRateId = ?'
         );
-        $deleteAdditionalChargeCustomerPricesStatement->execute([$additionalChargeRate->id()->value()]);
+        if (!$deleteAdditionalChargeCustomerPricesStatement->execute([$additionalChargeRate->id()->value()])) {
+            $errorInfo = json_encode($deleteAdditionalChargeCustomerPricesStatement->errorInfo());
+            throw new PDOException("Failed to delete additionalChargeRateCustomerPrices: {$errorInfo}");
+        }
+
     }
 
     /**
@@ -82,7 +92,6 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
      */
     private function insertAdditionalChargeCustomerSpecificPrices(AdditionalChargeRate $additionalChargeRate): void
     {
-
         $values     = [];
         $subQueries = [];
         foreach ($additionalChargeRate->specificCustomerPrices() as $specificCustomerPrice) {
@@ -96,11 +105,15 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
             );
             $subQueries[] = "(?,?,?)";
         }
-        if (!empty($values)) {
-            $query           = "insert into additionalChargeRateCustomerPrices(additionalChargeRateId, customerId, salePrice) values ";
-            $query           .= implode(',', $subQueries);
-            $insertStatement = $this->pdo->prepare($query);
-            $insertStatement->execute($values);
+        if (empty($values)) {
+            return;
+        }
+        $query           = "insert into additionalChargeRateCustomerPrices(additionalChargeRateId, customerId, salesPrice) values ";
+        $query           .= implode(',', $subQueries);
+        $insertStatement = $this->pdo->prepare($query);
+        if (!$insertStatement->execute($values)) {
+            $errorInfo = json_encode($insertStatement->errorInfo());
+            throw new PDOException("Failed to insert additionalChargeRateCustomerPrices: {$errorInfo}");
         }
     }
 
@@ -110,7 +123,7 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
             'select id,description,notes,salesPrice from additionalChargeRate'
         );
         if (!$additionalChargeRateStatement || !$additionalChargeRateStatement->execute()) {
-            $errorInfo = print_r($additionalChargeRateStatement->errorInfo());
+            $errorInfo = json_encode($additionalChargeRateStatement->errorInfo());
             throw new PDOException("Failed to retrieve AdditionalChargeRate: {$errorInfo}");
         }
         $toReturn = [];
@@ -135,7 +148,7 @@ class AdditionalChargeRatePDORepository implements AdditionalChargeRateRepositor
         if (!$data) {
             throw new AdditionalChargeRateNotFoundException($additionalChargeRateId);
         }
-        $data['specificCustomerPrices'] = [];
+        $data['specificCustomerPrices']  = [];
         $specificCustomerPricesStatement = $this->pdo->prepare(
             'select additionalChargeRateId, customerId, salePrice from additionalChargeRateCustomerPrices where additionalChargeRateId = ?'
         );
